% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R, R/volcano.view.R
\name{module.genelist}
\alias{module.genelist}
\title{Adds gene list module to dashboard}
\usage{
module.genelist(collection)

module.genelist(collection)
}
\arguments{
\item{collection}{A list where each element is a vector of genes. The elements should be named.}
}
\description{
Adds gene list module to dashboard
}
\examples{
 module.genelist(list(housekeeping = c('ALBUMIN', 'GAPDH'), special = c('VHL', 'IL6R')))
module.genelist(list(housekeeping = c('ALBUMIN', 'GAPDH'), special = c('VHL', 'IL6R')))
}
